# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Student APP API",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/classrooms": {
            "get": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "List all classrooms",
                "operationId": "df0e644fedc1687db78ae8f72d732531",
                "parameters": [
                    {
                        "name": "includeStudents",
                        "in": "query",
                        "description": "Whether to include the students in each classroom",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "includeDegrees",
                        "in": "query",
                        "description": "Whether to include the degree of each classroom",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of classrooms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClassroomResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Create a new classroom",
                "operationId": "feec68582ef509f7e15c310e562b0da4",
                "requestBody": {
                    "description": "Classroom data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "degree_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "degree_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The created classroom",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassroomResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/classrooms/{classroom}": {
            "get": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Retrieve a single classroom",
                "operationId": "1fd876f7e99c789ba4f62f574def8f52",
                "parameters": [
                    {
                        "name": "classroom",
                        "in": "path",
                        "description": "ID of the classroom to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "includeStudents",
                        "in": "query",
                        "description": "Whether to include the students in the classroom",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "includeDegree",
                        "in": "query",
                        "description": "Whether to include the degree for the classroom",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested classroom",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassroomResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Classroom not found"
                    }
                }
            }
        },
        "/api/v1/classrooms/{id}": {
            "put": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Update a classroom with PUT request",
                "operationId": "bdfa65d5f705165d93e6edf03f79be72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the classroom to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Classroom data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "degree_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Classroom updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassroomResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Classroom not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Delete a classroom",
                "operationId": "4df6cb056d8f3fd47eddc0d102b4c25a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the classroom to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "forceDelete",
                        "in": "query",
                        "description": "Whether to force delete the classroom and its related records",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Classroom deleted successfully"
                    },
                    "404": {
                        "description": "Classroom not found"
                    },
                    "422": {
                        "description": "Cannot delete classroom with attached students"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Update a classroom with PATCH request",
                "operationId": "41a34f9dc3e1de9e52cf7f690b6893ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the classroom to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Classroom data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "degree_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Classroom updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassroomResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Classroom not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClassroomResource": {},
...
        }
    }
}